fix: move edit slider
